Move PyreverseConfig to pylint.testutil
See : https://github.com/PyCQA/pylint/pull/5010#issuecomment-920291278
